+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # 2.10.0 - TBD
 - Updated the urls Competitive Companion runs on, it no longer runs on all urls but only on the urls that it has parsers for
-- Added a parser for Luogu Online Judge
+- Added a parser for Luogu
 - Fixed the UVa Online Judge parser
 
 # 2.9.2 - December 16th, 2019